+++ js/main.js
@@ -1,7 +1,6 @@
-import './setup.js';
 import {createPosts} from './data.js';
-import {createComments} from './data.js';
 
-console.log(
-  createPosts()
-);
+createPosts();
+
+
+console.log(createPosts());
