--- js/setup.js
@@ -1,29 +1,0 @@
-const DESCRIPTIONS = ['Посмотрите как мы катались на слонах, восторг!',
-  'Год активных тренировок в зале. Ну что, я красотка?',
-  'День рождения подруги',
-  'Я здесь в новом амплуа, не удивляйтесь',
-  'Лучшее время для маленькой сиесты',
-  'Собака лучше всех получилась на фото, впрочем, как обычно',
-  'Я уехал на край света, чтобы найти себя',
-  'Это мы, довольные и беззаботные',
-  'Не могла представить, что в Гватемале так красиво!',
-];
-
-const MESSAGES = ['Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают.',
-  'Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = ['Ира',
-  'Федя',
-  'Маша',
-  'Макс',
-  'Никита',
-  'Лейла',
-  'Юля'];
-
-const POSTS_COUNT = 25;
